Fix build on MacOSX: Use PRIx64. Add -Werror in certifier.mak

+++ src/sev-snp/snp_derive_key.h
@@ -9,7 +9,7 @@
 typedef unsigned char  uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned       uint32_t;
-typedef unsigned long  uint64_t;
+typedef unsigned long long  uint64_t;
 #endif
 
 #define FIELD_POLICY_SHIFT      (0)
